% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_addons.R
\name{anova.table}
\alias{anova.table}
\title{Creates an ANOVA table}
\usage{
\method{anova}{table}(fmla, data, ...)
}
\description{
Creates an ANOVA table
}
\examples{
my.table <- anova.table(education + age ~ treated | month_id, dt.data, out="tables/anova_table_1.tex")
}
